Updated Advanced Summon Behavior and Bug-fixes
I updated the Advanced Summon's behavior to now have a minimum freeze time and then after that freeze time be disabled once the NPC is too far away or gets hit by the player.

I also updated the sprites of the advanced summon, its recipe list, and bug-fixes some dead pixels.

I also adjusted the names of the items in the localization.

Co-Authored-By: Maksim Peterburgskiy <maksim.peterburgskiy@gmail.com>